follow-up to code review

+++ DQMServices/Components/plugins/DQMStoreStats.cc
@@ -296,7 +296,7 @@ int DQMStoreStats::calcstats(int mode = DQMStoreStats::considerAllME) {
             getEmptyMetric(it->getTH2I()->GetArray(), it->getNbinsX() + 2, it->getNbinsY() + 2, 0));
         curr->update(it->getNbinsX() * it->getNbinsY(),
                      getEmptyMetric(it->getTH2I()->GetArray(), it->getNbinsX() + 2, it->getNbinsY() + 2, 0),
-                     it->getNbinsX() * it->getNbinsY() * sizeof(double));
+                     it->getNbinsX() * it->getNbinsY() * sizeof(int));
         break;
       case MonitorElement::Kind::TPROFILE2D:
         currentSubfolder.AddBinsD(
